01_Scorpion/requirements.txt
pillow
exif
